Add option CFS_NATIVE for -march=native optimization for cfs and cfsdeps
* CFS_NATIVE has max 20% benefit for Xenon and much less on AMD and Mac
* fix compile errors and warnings for clang-12
* optimization flags are in CFS_OPT_FLAGS
* CFS_OPT_FLAGS is set in compiler.cmake and applied in
  FindCFSDEPS.cmake for CFSDEPS_*_FLAGS
* CFSDEPS_*_FLAGS could deserve some love for C++/C/Fortran differences
* Intel could deserve some love

+++ source/Optimization/Design/DesignSpace.cc
@@ -324,7 +324,7 @@ DesignSpace::DesignSpace(StdVector<RegionIdType>& reg_data, PtrParamNode pn, Ers
                 }
               }
 
-              de.SetDesign(random ? (((float) rand()/RAND_MAX) * (upper - lower) + lower) : initial);
+              de.SetDesign(random ? (((float) rand()/(float) RAND_MAX) * (upper - lower) + lower) : initial);
 
               data.Push_back(de);
               totalElements_.Push_back(&data.Last());
